spam: Continue if reading domain directory fails
Previously, if the virtual mailbox base contains a non-directory file,
the application would stop and gracefully terminate. Instead, ignore
this file and continue attempting to parse the remaining entries in the
virtual mailbox base.

I encountered this because the virtual mailbox base directory on my mail
server happened to contain a .bash_history file from a previous shell
session.